src/requirement.txt
anyio==3.4.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
click==8.0.3
colorama==0.4.4
distlib==0.3.3
fastapi==0.70.0
filelock==3.4.0
gitdb==4.0.9
GitPython==3.1.24
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
MarkupSafe==2.0.1
platformdirs==2.4.0
pydantic==1.8.2
python-dotenv==0.19.2
PyYAML==5.4.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.10.0
watchgod==0.7
websockets==10.1
